ai_agent/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.7
anthropic==0.7.7
requests==2.31.0
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1
websockets==12.0
streamlit==1.28.2
gradio==4.7.1
beautifulsoup4==4.12.2
selenium==4.15.2
scrapy==2.11.0
pandas==2.1.4
numpy==1.25.2
matplotlib==3.8.2
pillow==10.1.0
psutil==5.9.6
httpx==0.25.2
langchain==0.0.350
langchain-openai==0.0.2
chromadb==0.4.18
tiktoken==0.5.2
python-multipart==0.0.6
jinja2==3.1.2
markdown==3.5.1
PyYAML==6.0.1
rich==13.7.0
typer==0.9.0
click==8.1.7
schedule==1.2.0
watchdog==3.0.0
sqlalchemy==2.0.23
redis==5.0.1
celery==5.3.4